changing pricing from decimal to double type

+++ bitcoin_pricing.py
@@ -31,7 +31,6 @@ import sys
 import logging
 import datetime
 import MySQLdb
-from decimal import *
 
 LOG_CHANNEL = "bitcoin_pricing"
 LOG_FILE = "activation_errors.log"
@@ -111,13 +110,13 @@ class BitcoinPricing:
             elapsed_seconds = int(time_delta.total_seconds())
             if elapsed_seconds < API_REQUEST_PERIOD:
                 self.logger.debug("Found price of " + str(row[1]) + " for " + currency + " from " + str(elapsed_seconds) + " seconds ago.")
-                return row[1]
+                return double(row[1])
             else:
                 self.fetch_bitcoin_pricing()
                 x.execute("SELECT * FROM bitcoin_pricing WHERE currency_symbol='%s' ORDER BY serial DESC LIMIT 1;" % currency)
                 row = x.fetchone()
                 self.logger.debug("Fetched price of " + str(row[1]) + " for " + currency + " from API server.")
-                return row[1]
+                return double(row[1])
 
         except MySQLdb.Error as e:
             self.logger.error("MySQLdb raised exception: %s",str(e))
